Add PositionIdentifier to Inspection (#36)
* initial commit

* need to allow Generalizedidentifier

* fixing a typo

* initial pass, but untested and missing SectionMember

* adding additional describe

* updating exports

* fixed incorrect uppercase

* adding framework for positionidentifier testing

* split up tests

* replacing maybeAstNode with isClosed

* initial passing tests

* added some ParserContext tests

* added failing test for section

* new test, fixed off-by-one error
