+++ src/inspection/index.ts
@@ -3,3 +3,4 @@
 
 export * from "./node";
 export * from "./inspection";
+export * from "./positionIdentifier";
